+++ MatrixSDK/Data/MXRoomState.m
@@ -20,6 +20,7 @@
 
 #import "MXSession.h"
 #import "MXTools.h"
+#import "MXCallManager.h"
 
 @interface MXRoomState ()
 {
@@ -55,6 +56,11 @@
      The key is the 3pid invite token.
      */
     NSMutableDictionary<NSString*, MXRoomMember*> *membersWithThirdPartyInviteTokenCache;
+
+    /**
+     The cache for the conference user id.
+     */
+    NSString *conferenceUserId;
 }
 @end
 
@@ -162,6 +168,11 @@
 - (NSArray *)members
 {
     return [members allValues];
+}
+
+- (NSArray<MXRoomMember *> *)joinedMembers
+{
+    return [self membersWithMembership:MXMembershipJoin];
 }
 
 - (NSArray<MXRoomThirdPartyInvite *> *)thirdPartyInvites
@@ -425,6 +436,7 @@
     return result;
 }
 
+
 #pragma mark - State events handling
 - (void)handleStateEvent:(MXEvent*)event
 {
@@ -468,6 +480,11 @@
                 {
                     [self handleStateEvent:inviteRoomStateEvent];
                 }
+            }
+            else if (_isLive && self.membership == MXMembershipJoin && members.count > 2 && [roomMember.userId isEqualToString:self.conferenceUserId])
+            {
+                // Forward the change of the conference user membership to the call manager 
+                [mxSession.callManager handleConferenceUserUpdate:roomMember inRoom:_roomId];
             }
 
             break;
@@ -639,6 +656,125 @@
     return powerLevel;
 }
 
+- (NSArray<MXRoomMember*>*)membersWithMembership:(MXMembership)theMembership
+{
+    NSMutableArray *membersWithMembership = [NSMutableArray array];
+    for (MXRoomMember *roomMember in members.allValues)
+    {
+        if (roomMember.membership == theMembership)
+        {
+            [membersWithMembership addObject:roomMember];
+        }
+    }
+    return membersWithMembership;
+}
+
+
+# pragma mark - Conference call
+- (BOOL)isOngoingConferenceCall
+{
+    BOOL isOngoingConferenceCall = NO;
+
+    MXRoomMember *conferenceUserMember = [self memberWithUserId:self.conferenceUserId];
+    if (conferenceUserMember)
+    {
+        isOngoingConferenceCall = (conferenceUserMember.membership == MXMembershipJoin);
+    }
+
+    return isOngoingConferenceCall;
+}
+
+- (BOOL)isConferenceUserRoom
+{
+    BOOL isConferenceUserRoom = NO;
+
+    // A conference user room is a 1:1 room with a conference user
+    if (members.count == 2)
+    {
+        for (NSString *memberUserId in members)
+        {
+            if ([MXCallManager isConferenceUser:memberUserId])
+            {
+                isConferenceUserRoom = YES;
+                break;
+            }
+        }
+    }
+
+    return isConferenceUserRoom;
+}
+
+- (NSString *)conferenceUserId
+{
+    if (!conferenceUserId)
+    {
+        conferenceUserId = [MXCallManager conferenceUserIdForRoom:_roomId];
+    }
+    return conferenceUserId;
+}
+
+- (NSArray<MXRoomMember *> *)membersWithoutConferenceUser
+{
+    NSArray *membersWithoutConferenceUser;
+
+    if (self.isConferenceUserRoom)
+    {
+        // Show everyone in a 1:1 room with a conference user
+        membersWithoutConferenceUser = self.members;
+    }
+    else if (![self memberWithUserId:self.conferenceUserId])
+    {
+        // There is no conference user. No need to filter
+        membersWithoutConferenceUser = self.members;
+    }
+    else
+    {
+        // Filter the conference user from the list
+        NSMutableDictionary *membersWithoutConferenceUserDict = [NSMutableDictionary dictionaryWithDictionary:members];
+        [membersWithoutConferenceUserDict removeObjectForKey:self.conferenceUserId];
+        membersWithoutConferenceUser = membersWithoutConferenceUserDict.allValues;
+    }
+
+    return membersWithoutConferenceUser;
+}
+
+- (NSArray<MXRoomMember *> *)membersWithMembership:(MXMembership)theMembership includeConferenceUser:(BOOL)includeConferenceUser
+{
+    NSArray *membersWithMembership;
+
+    if (includeConferenceUser || self.isConferenceUserRoom)
+    {
+        // Show everyone in a 1:1 room with a conference user
+        membersWithMembership = [self membersWithMembership:theMembership];
+    }
+    else
+    {
+        MXRoomMember *conferenceUserMember = [self memberWithUserId:self.conferenceUserId];
+        if (!conferenceUserMember || conferenceUserMember.membership != theMembership)
+        {
+            // The conference user is not in list of members with the passed  membership
+            membersWithMembership = [self membersWithMembership:theMembership];
+        }
+        else
+        {
+            NSMutableDictionary *membersWithMembershipDict = [NSMutableDictionary dictionaryWithCapacity:members.count];
+            for (MXRoomMember *roomMember in members.allValues)
+            {
+                if (roomMember.membership == theMembership)
+                {
+                    membersWithMembershipDict[roomMember.userId] = roomMember;
+                }
+            }
+
+            [membersWithMembershipDict removeObjectForKey:self.conferenceUserId];
+            membersWithMembership = membersWithMembershipDict.allValues;
+        }
+    }
+
+    return membersWithMembership;
+}
+
+
 #pragma mark - NSCopying
 - (id)copyWithZone:(NSZone *)zone
 {
@@ -667,6 +803,11 @@
     
     stateCopy->powerLevels = [powerLevels copy];
     stateCopy->maxPowerLevel = maxPowerLevel;
+
+    if (conferenceUserId)
+    {
+        stateCopy->conferenceUserId = [conferenceUserId copyWithZone:zone];
+    }
 
     return stateCopy;
 }