+++ MatrixSDK/VoIP/MXCall.h
@@ -44,9 +44,15 @@ typedef enum : NSUInteger
     MXCallStateConnected,
     MXCallStateEnded,
 
-    MXCallStateInviteExpired
+    MXCallStateInviteExpired,
+    MXCallStateAnsweredElseWhere
 } MXCallState;
 
+/**
+ Posted when a `MXCall` object has changed its state.
+ The notification object is the `MXKCall` object representing the call.
+ */
+extern NSString *const kMXCallStateDidChange;
 
 @protocol MXCallDelegate;
 
@@ -63,6 +69,16 @@ typedef enum : NSUInteger
  @return the newly created MXCall instance.
  */
 - (instancetype)initWithRoomId:(NSString*)roomId andCallManager:(MXCallManager*)callManager;
+
+/**
+ Create a `MXCall` instance in order to place a call using a conference server.
+
+ @param roomId the id of the room where to place the call.
+ @param callSignalingRoomId the id of the room where call signaling is managed with the conference server.
+ @param callManager the manager of all MXCall objects.
+ @return the newly created MXCall instance.
+ */
+- (instancetype)initWithRoomId:(NSString*)roomId callSignalingRoomId:(NSString*)callSignalingRoomId andCallManager:(MXCallManager*)callManager;
 
 /**
  Handle call event.
@@ -96,9 +112,21 @@ typedef enum : NSUInteger
 @property (readonly, nonatomic) MXRoom *room;
 
 /**
+ The room where the signaling of the call is managed.
+ It is same value as 'room' in case of 1:1 call.
+ It is a private room with the conference user in case of conference call.
+ */
+@property (readonly, nonatomic) MXRoom *callSignalingRoom;
+
+/**
  The id of the call.
  */
 @property (readonly, nonatomic) NSString *callId;
+
+/**
+ Flag indicating this is a conference call;
+ */
+@property (readonly, nonatomic) BOOL isConferenceCall;
 
 /**
  Flag indicating if this is an incoming call.
@@ -135,6 +163,24 @@ typedef enum : NSUInteger
  on the other peer device.
  */
 @property (nonatomic) UIDeviceOrientation selfOrientation;
+
+/**
+ Mute state of the audio.
+ */
+@property (nonatomic) BOOL audioMuted;
+
+/**
+ Mute state of the video.
+ */
+@property (nonatomic) BOOL videoMuted;
+
+/**
+ If YES, the inbound audio is sent to the main speaker, else it is routed to the
+ ear speaker.
+ By default, video calls are routed to the main speaker and voice call, to the ear
+ speaker.
+ */
+@property (nonatomic) BOOL audioToSpeaker;
 
 /**
  The call duration in milliseconds.