+++ MatrixSDK/VoIP/MXCall.m
@@ -19,6 +19,9 @@
 #import "MXSession.h"
 #import "MXCallStackCall.h"
 
+#pragma mark - Constants definitions
+NSString *const kMXCallStateDidChange = @"kMXCallStateDidChange";
+
 @interface MXCall ()
 {
     /**
@@ -50,25 +53,47 @@
      Timer to expire an invite.
      */
     NSTimer *inviteExpirationTimer;
+
+    /**
+     A queue of gathered local ICE candidates waiting to be sent to the other peer.
+     */
+    NSMutableArray<NSDictionary*> *localICECandidates;
+
+    /**
+     Timer for sending local ICE candidates.
+     */
+    NSTimer *localIceGatheringTimer;
 }
 
 @end
 
 @implementation MXCall
 
-- (instancetype)initWithRoomId:(NSString *)roomId andCallManager:(MXCallManager *)callManager2
+- (instancetype)initWithRoomId:(NSString *)roomId andCallManager:(MXCallManager *)theCallManager
+{
+    // For 1:1 call, use the room as the call signaling room
+    return [self initWithRoomId:roomId callSignalingRoomId:roomId andCallManager:theCallManager];
+}
+
+- (instancetype)initWithRoomId:(NSString*)roomId callSignalingRoomId:(NSString*)callSignalingRoomId andCallManager:(MXCallManager*)theCallManager;
 {
     self = [super init];
     if (self)
     {
-        callManager = callManager2;
+        callManager = theCallManager;
 
         _room = [callManager.mxSession roomWithRoomId:roomId];
+        _callSignalingRoom = [callManager.mxSession roomWithRoomId:callSignalingRoomId];
 
         _callId = [[NSUUID UUID] UUIDString];
         _callerId = callManager.mxSession.myUser.userId;
 
         _state = MXCallStateFledgling;
+
+        // Consider we are using a conference call when there are more than 2 users
+        _isConferenceCall = (2 <= _room.state.joinedMembers.count);
+
+        localICECandidates = [NSMutableArray array];
 
         callStackCall = [callManager.callStack createCall];
         if (nil == callStackCall)
@@ -103,7 +128,7 @@
         {
             callInviteEventContent = [MXCallInviteEventContent modelFromJSON:event.content];
 
-            if (NO == [event.sender isEqualToString:_room.mxSession.myUser.userId])
+            if (NO == [event.sender isEqualToString:_callSignalingRoom.mxSession.myUser.userId])
             {
                 // Incoming call
 
@@ -113,6 +138,9 @@
 
                 // Store if it is voice or video call
                 _isVideoCall = callInviteEventContent.isVideoCall;
+
+                // Set up the default audio route for this call type
+                callStackCall.audioToSpeaker = _isVideoCall;
                 
                 [callStackCall startCapturingMediaWithVideo:self.isVideoCall success:^{
                     [callStackCall handleOffer:callInviteEventContent.offer.sdp];
@@ -141,26 +169,36 @@
 
         case MXEventTypeCallAnswer:
         {
-            // MXCall receives this event only when it placed a call
-            MXCallAnswerEventContent *content = [MXCallAnswerEventContent modelFromJSON:event.content];
-
-            // The peer accepted our outgoing call
-            if (inviteExpirationTimer)
+            // Listen to answer event only for call we are making, not receiving
+            if (NO == _isIncoming)
             {
-                [inviteExpirationTimer invalidate];
-                inviteExpirationTimer = nil;
+                // MXCall receives this event only when it placed a call
+                MXCallAnswerEventContent *content = [MXCallAnswerEventContent modelFromJSON:event.content];
+
+                // The peer accepted our outgoing call
+                if (inviteExpirationTimer)
+                {
+                    [inviteExpirationTimer invalidate];
+                    inviteExpirationTimer = nil;
+                }
+
+                // Let's the stack finalise the connection
+                [callStackCall handleAnswer:content.answer.sdp success:^{
+
+                    // Call is up
+                    [self setState:MXCallStateConnected reason:event];
+
+                } failure:^(NSError *error) {
+                    NSLog(@"[MXCall] handleCallEvent: ERROR: Cannot send handle answer. Error: %@\nEvent: %@", error, event);
+                    [self didEncounterError:error];
+                }];
             }
-
-            // Let's the stack finalise the connection
-            [callStackCall handleAnswer:content.answer.sdp success:^{
-
-                // Call is up
-                [self setState:MXCallStateConnected reason:event];
-
-            } failure:^(NSError *error) {
-                NSLog(@"[MXCall] handleCallEvent: ERROR: Cannot send handle answer. Error: %@\nEvent: %@", error, event);
-                [self didEncounterError:error];
-            }];
+            else if (_state == MXCallStateRinging)
+            {
+                // Else this event means that the call has been answered by the user from
+                // another device
+                [self onCallAnsweredElsewhere];
+            }
             break;
         }
 
@@ -175,12 +213,15 @@
 
         case MXEventTypeCallCandidates:
         {
-            MXCallCandidatesEventContent *content = [MXCallCandidatesEventContent modelFromJSON:event.content];
-
-            NSLog(@"[MXCall] handleCallCandidates: %@", content.candidates);
-            for (MXCallCandidate *canditate in content.candidates)
+            if (NO == [event.sender isEqualToString:_callSignalingRoom.mxSession.myUser.userId])
             {
-                [callStackCall handleRemoteCandidate:canditate.JSONDictionary];
+                MXCallCandidatesEventContent *content = [MXCallCandidatesEventContent modelFromJSON:event.content];
+
+                NSLog(@"[MXCall] handleCallCandidates: %@", content.candidates);
+                for (MXCallCandidate *canditate in content.candidates)
+                {
+                    [callStackCall handleRemoteCandidate:canditate.JSONDictionary];
+                }
             }
             break;
         }
@@ -202,6 +243,9 @@
 
     [self setState:MXCallStateWaitLocalMedia reason:nil];
 
+    // Set up the default audio route for this call type
+    callStackCall.audioToSpeaker = video;
+
     [callStackCall startCapturingMediaWithVideo:video success:^() {
 
         [callStackCall createOffer:^(NSString *sdp) {
@@ -220,7 +264,7 @@
                                       @"version": @(0),
                                       @"lifetime": @(callManager.inviteLifetime)
                                       };
-            [_room sendEventOfType:kMXEventTypeStringCallInvite content:content success:^(NSString *eventId) {
+            [_callSignalingRoom sendEventOfType:kMXEventTypeStringCallInvite content:content success:^(NSString *eventId) {
 
                 [self setState:MXCallStateInviteSent reason:nil];
 
@@ -272,8 +316,9 @@
                                               },
                                       @"version": @(0),
                                       };
-            [_room sendEventOfType:kMXEventTypeStringCallAnswer content:content success:^(NSString *eventId) {
-                
+            [_callSignalingRoom sendEventOfType:kMXEventTypeStringCallAnswer content:content success:^(NSString *eventId) {
+
+                // @TODO: This is false
                 [self setState:MXCallStateConnected reason:nil];
                 
             } failure:^(NSError *error) {
@@ -303,7 +348,7 @@
                                   @"call_id": _callId,
                                   @"version": @(0)
                                   };
-        [_room sendEventOfType:kMXEventTypeStringCallHangup content:content success:nil failure:^(NSError *error) {
+        [_callSignalingRoom sendEventOfType:kMXEventTypeStringCallHangup content:content success:nil failure:^(NSError *error) {
             NSLog(@"[MXCall] hangup: ERROR: Cannot send m.call.hangup event. Error: %@", error);
             [self didEncounterError:error];
         }];
@@ -332,6 +377,9 @@
     {
         [_delegate call:self stateDidChange:_state reason:event];
     }
+
+    // Broadcast the new call state
+    [[NSNotificationCenter defaultCenter] postNotificationName:kMXCallStateDidChange object:self userInfo:nil];
 }
 
 - (void)setSelfVideoView:(UIView *)selfVideoView
@@ -365,6 +413,36 @@
     }
 }
 
+- (BOOL)audioMuted
+{
+    return callStackCall.audioMuted;
+}
+
+- (void)setAudioMuted:(BOOL)audioMuted
+{
+    callStackCall.audioMuted = audioMuted;
+}
+
+- (BOOL)videoMuted
+{
+    return callStackCall.videoMuted;
+}
+
+- (void)setVideoMuted:(BOOL)videoMuted
+{
+    callStackCall.videoMuted = videoMuted;
+}
+
+- (BOOL)audioToSpeaker
+{
+    return callStackCall.audioToSpeaker;
+}
+
+- (void)setAudioToSpeaker:(BOOL)audioToSpeaker
+{
+    callStackCall.audioToSpeaker = audioToSpeaker;
+}
+
 - (NSUInteger)duration
 {
     NSUInteger duration = 0;
@@ -382,7 +460,47 @@
 
 
 #pragma mark - MXCallStackCallDelegate
--(void)callStackCall:(id<MXCallStackCall>)callStackCall onError:(NSError *)error
+- (void)callStackCall:(id<MXCallStackCall>)callStackCall onICECandidateWithSdpMid:(NSString *)sdpMid sdpMLineIndex:(NSInteger)sdpMLineIndex candidate:(NSString *)candidate
+{
+    // Candidates are sent in a special way because we try to amalgamate
+    // them into one message
+    // No need for locking data as everything is running on the main thread
+    [localICECandidates addObject:@{
+                                    @"sdpMid": sdpMid,
+                                    @"sdpMLineIndex": @(sdpMLineIndex),
+                                    @"candidate":candidate
+                                    }
+     ];
+
+    // Send candidates every 100ms max. This value gives enough time to the underlaying call stack
+    // to gather several ICE candidates
+    [localIceGatheringTimer invalidate];
+    localIceGatheringTimer = [NSTimer scheduledTimerWithTimeInterval:0.1 target:self selector:@selector(sendLocalIceCandidates) userInfo:self repeats:NO];
+}
+
+- (void)sendLocalIceCandidates
+{
+    localIceGatheringTimer = nil;
+
+    if (localICECandidates.count)
+    {
+        NSLog(@"MXCall] onICECandidate: Send %tu candidates", localICECandidates.count);
+
+        NSDictionary *content = @{
+                                  @"version": @(0),
+                                  @"call_id": _callId,
+                                  @"candidates": localICECandidates
+                                  };
+
+        [_callSignalingRoom sendEventOfType:kMXEventTypeStringCallCandidates content:content success:nil failure:^(NSError *error) {
+            NSLog(@"[MXCall] onICECandidate: Warning: Cannot send m.call.candidates event. Error: %@", error);
+        }];
+
+        [localICECandidates removeAllObjects];
+    }
+}
+
+- (void)callStackCall:(id<MXCallStackCall>)callStackCall onError:(NSError *)error
 {
     NSLog(@"[MXCall] callStackCall didEncounterError: %@", error);
     [self didEncounterError:error];
@@ -397,6 +515,10 @@
         [inviteExpirationTimer invalidate];
         inviteExpirationTimer = nil;
     }
+
+    // Do not refresh TURN servers config anymore
+    [localIceGatheringTimer invalidate];
+    localIceGatheringTimer = nil;
 
     // Terminate the call at the stack level
     [callStackCall end];
@@ -433,6 +555,18 @@
         // The call manager can now ignore this call
         [callManager removeCall:self];
     }
+}
+
+- (void)onCallAnsweredElsewhere
+{
+    // Send the notif that the call has been answered from another device to the app
+    [self setState:MXCallStateAnsweredElseWhere reason:nil];
+
+    // And set the final state: MXCallStateEnded
+    [self setState:MXCallStateEnded reason:nil];
+
+    // The call manager can now ignore this call
+    [callManager removeCall:self];
 }
 
 @end