+++ CHANGES.rst
@@ -1,5 +1,33 @@
+Changes in Matrix iOS SDK in 0.6.12 (2016-08-01)
+================================================
+
+Improvements:
+ * MXCallManager: Better handle call invites when the app resumes.
+ * MXCall: Improve the sending of local ICE candidates to avoid HTTP 429(Too Many Requests) response
+ * MXCall: Added the audioToSpeaker property to choose between the main and the ear speaker.
+ * MXRoomState: Added the joinedMembers property.
+ * MXLogger: Added the isMainThread information in crash logs.
+ 
+Bug fixes:
+ * MXJingleCallStackCall: Added sanity check on creation of RTCICEServer objects as crashes have been reported.
+
+Changes in Matrix iOS SDK in 0.6.11 (2016-07-26)
+================================================
+
+Improvements:
+ * MXCall: Added audioMuted and videoMuted properties.
+ * Call: the SDK is now able to send local ICE candidates.
+ * Integration of libjingle/PeerConnection call stack (see MXJingleCall).
+ 
+Bug fixes:
+ * MXCallManager: Do not show the call screen when the call is initiated by the same user but from another device.
+ * MXCallManager: Hide the call screen when the user answers an incoming call from another device.
+
+Breaks:
+ * MXCallStackCall: two new properties (audioMuted and videoMuted) and one new delegate method (onICECandidateWithSdpMid).
+
 Changes in Matrix iOS SDK in 0.6.10 (2016-07-15)
-===============================================
+================================================
 
 Improvements:
  * MXRestClient: Add API to add/remove a room alias.